requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-api-python-client==2.111.0
google-auth==2.25.2
python-dotenv==1.0.0
APScheduler==3.10.4
gunicorn==21.2.0
requests==2.31.0
pytz==2024.1
psycopg2-binary==2.9.9
